extraction - cloning repo and removing all but subset to extract
Part of process of breaking out the assets content (a single
skeleton/template at time of action) into its own module and using it
as a git submodule.

This will keep the git history that impacted the template repo.  It will
also keep that of the cleared surrounding content, but I consider this
acceptable.  Alternative would be using specialty tools like
`git-filter-repo`, with concomittent unknowns introduced.